feat(Sidebar): add optional icons for groups and enhance mixed content support
